+++ CHANGES.rst
@@ -9,7 +9,7 @@ date: FIXME
 - Start supporting Python v3.14.
 - Feature: Allow to configure histogram step size and scale from code.
 - Cosmetic: Show *leftmost* part of long offset/gain/gamma values by default.
-- Visual: Add configuration setting 'color_scheme' (value: null / 'Dark' / 'Light').
+- Visual: Add configuration setting 'color_scheme' (value: null / 'Auto' / 'Dark' / 'Light').
 - Visual: Add separate setting 'image background dark' (background color in dark mode).
 - Visual: Draw more widgets with suitable colors in dark mode.
 - ...